requirements.txt
langchain
streamlit
google-generativeai
PyPDF2   # Read  pdf
faiss-cpu  # vector db
langchain_google_genai
python-dotenv
